Get Python specific templates into src/controller/python
