+++ scripts/tools/zap_regen_all.py
@@ -37,6 +37,7 @@ def getGlobalTemplatesTargets():
 def getSpecificTemplatesTargets():
     targets = []
     targets.append(['examples/chip-tool/chip-tool.zap', '-t', 'examples/chip-tool/templates/templates.json'])
+    targets.append(['src/controller/python/local-clusters.zap', '-t', 'src/controller/python/templates/templates.json'])
     targets.append(['src/controller/controller-clusters.zap', '-t', 'src/app/zap-templates/chip-templates.json'])
     targets.append(['src/darwin/Framework/CHIP/chip-tool.zap', '-t', 'src/darwin/Framework/CHIP/templates/templates.json'])
     return targets
